+++ predicate/src/predicate.rs
@@ -188,7 +188,7 @@ impl Predicate {
     }
 
     /// Merge the given delete predicates into this select predicate.
-    /// Since we wan to eliminate data filtered by the delete predicates,
+    /// Since we want to eliminate data filtered by the delete predicates,
     /// they are first converted into their negated form: NOT(delete_predicate)
     /// then added/merged into the selection one
     pub fn merge_delete_predicates<S>(&mut self, delete_predicates: &[S])
@@ -197,7 +197,7 @@ impl Predicate {
     {
         // Create a list of disjunctive negated expressions.
         // Example: there are two deletes as follows (note that time_range is stored separated in the Predicate
-        //  but we need to put it together with the exprs hee)
+        //  but we need to put it together with the exprs here)
         //   . Delete_1: WHERE city != "Boston"  AND temp = 70  AND time_range in [10, 30)
         //   . Delete 2: WHERE state = "NY" AND route != "I90" AND time_range in [20, 50)
         // The negated list will be "NOT(Delete_1)", NOT(Delete_2)" which means
